view favorites

+++ packages/twenty-front/src/modules/favorites/components/CurrentWorkspaceMemberFavorites.tsx
@@ -1,10 +1,7 @@
 import { useTheme } from '@emotion/react';
-import styled from '@emotion/styled';
 import { useState } from 'react';
 import { useLocation } from 'react-router-dom';
 import {
-  Avatar,
-  AvatarType,
   IconDotsVertical,
   IconFolder,
   IconHeartOff,
@@ -12,13 +9,17 @@ import {
   IconTrash,
 } from 'twenty-ui';
 
+import { FavoriteIcon } from '@/favorites/components/FavoriteIcon';
 import { FavoriteFolderHotkeyScope } from '@/favorites/constants/FavoriteFolderRightIconDropdownHotkeyScope';
 import { useFavoriteFolders } from '@/favorites/hooks/useFavoriteFolders';
 import { useFavorites } from '@/favorites/hooks/useFavorites';
+import { ProcessedFavorite } from '@/favorites/utils/sortFavorites';
 import { DraggableItem } from '@/ui/layout/draggable-list/components/DraggableItem';
 import { DraggableList } from '@/ui/layout/draggable-list/components/DraggableList';
 import { Dropdown } from '@/ui/layout/dropdown/components/Dropdown';
+import { DropdownMenuItemsContainer } from '@/ui/layout/dropdown/components/DropdownMenuItemsContainer';
 import { useDropdown } from '@/ui/layout/dropdown/hooks/useDropdown';
+import { ConfirmationModal } from '@/ui/layout/modal/components/ConfirmationModal';
 import { MenuItem } from '@/ui/navigation/menu-item/components/MenuItem';
 import { NavigationDrawerInput } from '@/ui/navigation/navigation-drawer/components/NavigationDrawerInput';
 import { NavigationDrawerItem } from '@/ui/navigation/navigation-drawer/components/NavigationDrawerItem';
@@ -26,28 +27,11 @@ import { NavigationDrawerItemsCollapsedContainer } from '@/ui/navigation/navigat
 import { NavigationDrawerSubItem } from '@/ui/navigation/navigation-drawer/components/NavigationDrawerSubItem';
 import { getNavigationSubItemState } from '@/ui/navigation/navigation-drawer/utils/getNavigationSubItemState';
 
-const StyledAvatar = styled(Avatar)`
-  :hover {
-    cursor: grab;
-  }
-`;
-
-const StyledDropdownContainer = styled.div`
-  padding: ${({ theme }) => theme.spacing(1)};
-`;
-
 type CurrentWorkspaceMemberFavoritesProps = {
   folder: {
     folderId: string;
     folderName: string;
-    favorites: Array<{
-      id: string;
-      labelIdentifier: string;
-      avatarUrl: string;
-      avatarType: AvatarType;
-      link: string;
-      recordId: string;
-    }>;
+    favorites: ProcessedFavorite[];
   };
   isGroup: boolean;
   isOpen: boolean;
@@ -61,15 +45,21 @@ export const CurrentWorkspaceMemberFavorites = ({
   onToggle,
 }: CurrentWorkspaceMemberFavoritesProps) => {
   const currentPath = useLocation().pathname;
+  const currentPathView = useLocation().pathname + useLocation().search;
+
   const theme = useTheme();
   const [isRenaming, setIsRenaming] = useState(false);
   const [folderName, setFolderName] = useState(folder.folderName);
+  const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
+
   const { renameFolder, deleteFolder } = useFavoriteFolders();
   const { closeDropdown, isDropdownOpen } = useDropdown(
     `favorite-folder-${folder.folderId}`,
   );
-  const selectedFavoriteIndex = folder.favorites.findIndex(
-    (favorite) => favorite.link === currentPath,
+  const selectedFavoriteIndex = folder.favorites.findIndex((favorite) =>
+    favorite.objectNameSingular === 'view'
+      ? favorite.link === currentPathView
+      : favorite.link === currentPath,
   );
   const { deleteFavorite, handleReorderFavorite } = useFavorites();
 
@@ -102,8 +92,20 @@ export const CurrentWorkspaceMemberFavorites = ({
   };
 
   const handleDelete = async () => {
-    await deleteFolder(folder.folderId);
+    if (folder.favorites.length > 0) {
+      setIsDeleteModalOpen(true);
+      closeDropdown();
+    } else {
+      await deleteFolder(folder.folderId);
+      closeDropdown();
+    }
   };
+
+  const handleConfirmDelete = async () => {
+    await deleteFolder(folder.folderId);
+    setIsDeleteModalOpen(false);
+  };
+
   const rightOptions = (
     <Dropdown
       dropdownId={`favorite-folder-${folder.folderId}`}
@@ -120,60 +122,56 @@ export const CurrentWorkspaceMemberFavorites = ({
       dropdownPlacement="right"
       dropdownOffset={{ y: -15 }}
       dropdownComponents={
-        <StyledDropdownContainer>
+        <DropdownMenuItemsContainer>
           <MenuItem
             LeftIcon={IconPencil}
             onClick={() => {
               setIsRenaming(true);
               closeDropdown();
             }}
-            accent={'default'}
-            text={'Rename'}
+            accent="default"
+            text="Rename"
           />
           <MenuItem
             LeftIcon={IconTrash}
-            onClick={() => {
-              handleDelete();
-              closeDropdown();
-            }}
-            accent={'danger'}
-            text={'Delete'}
+            onClick={handleDelete}
+            accent="danger"
+            text="Delete"
           />
-        </StyledDropdownContainer>
+        </DropdownMenuItemsContainer>
       }
     />
   );
 
   return (
-    <NavigationDrawerItemsCollapsedContainer
-      key={folder.folderId}
-      isGroup={isGroup}
-    >
-      {isRenaming ? (
-        <NavigationDrawerInput
-          Icon={IconFolder}
-          value={folderName}
-          onChange={setFolderName}
-          onSubmit={handleSubmitRename}
-          onCancel={handleCancelRename}
-          onClickOutside={handleClickOutside}
-          hotkeyScope="favorites-folder-input"
-        />
-      ) : (
-        <NavigationDrawerItem
-          key={folder.folderId}
-          label={folder.folderName}
-          Icon={IconFolder}
-          onClick={() => onToggle(folder.folderId)}
-          active={isOpen}
-          rightOptions={rightOptions}
-          className="navigation-drawer-item"
-          isDropdownOpen={isDropdownOpen}
-        />
-      )}
+    <>
+      <NavigationDrawerItemsCollapsedContainer
+        key={folder.folderId}
+        isGroup={isGroup}
+      >
+        {isRenaming ? (
+          <NavigationDrawerInput
+            Icon={IconFolder}
+            value={folderName}
+            onChange={setFolderName}
+            onSubmit={handleSubmitRename}
+            onCancel={handleCancelRename}
+            onClickOutside={handleClickOutside}
+            hotkeyScope="favorites-folder-input"
+          />
+        ) : (
+          <NavigationDrawerItem
+            key={folder.folderId}
+            label={folder.folderName}
+            Icon={IconFolder}
+            onClick={() => onToggle(folder.folderId)}
+            rightOptions={rightOptions}
+            className="navigation-drawer-item"
+            isDropdownOpen={isDropdownOpen}
+          />
+        )}
 
-      {isOpen && (
-        <NavigationDrawerItemsCollapsedContainer isGroup={isGroup}>
+        {isOpen && (
           <DraggableList
             onDragEnd={handleReorderFavorite}
             draggableItems={
@@ -188,17 +186,13 @@ export const CurrentWorkspaceMemberFavorites = ({
                         key={favorite.id}
                         className="navigation-drawer-sub-item"
                         label={favorite.labelIdentifier}
-                        Icon={() => (
-                          <StyledAvatar
-                            placeholderColorSeed={favorite.recordId}
-                            avatarUrl={favorite.avatarUrl}
-                            type={favorite.avatarType}
-                            placeholder={favorite.labelIdentifier}
-                            className="fav-avatar"
-                          />
-                        )}
+                        Icon={() => <FavoriteIcon favorite={favorite} />}
                         to={favorite.link}
-                        active={favorite.link === currentPath}
+                        active={
+                          favorite.objectNameSingular === 'view'
+                            ? favorite.link === currentPathView
+                            : favorite.link === currentPath
+                        }
                         subItemState={getNavigationSubItemState({
                           index,
                           arrayLength: subItemArrayLength,
@@ -207,7 +201,7 @@ export const CurrentWorkspaceMemberFavorites = ({
                         rightOptions={
                           <IconHeartOff
                             size={theme.icon.size.sm}
-                            color={theme.color.red}
+                            color={theme.color.gray50}
                             onClick={() => deleteFavorite(favorite.id)}
                           />
                         }
@@ -218,8 +212,19 @@ export const CurrentWorkspaceMemberFavorites = ({
               </>
             }
           />
-        </NavigationDrawerItemsCollapsedContainer>
-      )}
-    </NavigationDrawerItemsCollapsedContainer>
+        )}
+      </NavigationDrawerItemsCollapsedContainer>
+
+      <ConfirmationModal
+        isOpen={isDeleteModalOpen}
+        setIsOpen={setIsDeleteModalOpen}
+        confirmationValue="yes"
+        confirmationPlaceholder="yes"
+        title={`Remove ${folder.favorites.length} favorites?`}
+        subtitle={`This action will delete this favorite folder and all ${folder.favorites.length} favorites inside. Do you want to continue?`}
+        onConfirmClick={handleConfirmDelete}
+        deleteButtonText="Delete Folder"
+      />
+    </>
   );
 };
